+++ DTO_UIL_BOL/DTO_UIL_BOL.cs
@@ -56,8 +56,8 @@ namespace DTO_UIL_BOL
         public int SalaireHebdomadaireDto { get; set; }
         public int? TailleDto { get; set; }
         public int? PoidDto { get; set; }
-        public int PrixMinDto { get; set; }
-        public int? PrixMaxDto { get; set; }
+        public double PrixMinDto { get; set; }
+        public double? PrixMaxDto { get; set; }
         public string PositionDto { get; set; }
         public string NomEquipeDto { get; set; }
         public string NomPaysDto { get; set; }
@@ -69,7 +69,7 @@ namespace DTO_UIL_BOL
     public class OffreDTO
     {
         public int IdJoueurDto { get; set; }
-        public int PrixOffreDto { get; set; }
+        public double PrixOffreDto { get; set; }
     }
     #endregion
 }
